Added weapon abstract class, aiming functionality
An abstract class RangedWeapon now defines shared behavior for weapon types, reading from a WeaponData scriptable object. The weapon has support for auto, semiauto, bolt(which may need to be replaced with the vanquish implementation, i.e a semi auto weapon with a 1 round magazine, and beam, which is.... uhhhh.... hitscan auto with a very high fire rate and no recoil? Might need to axe that one too. Gizmos now represent the reticle and the current bloom.